% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMyFifaCloud.R
\name{buildMyFifaCloud}
\alias{buildMyFifaCloud}
\title{Build FIFA-themed word cloud from a Document Term Matrix}
\usage{
buildMyFifaCloud(dtm)
}
\arguments{
\item{dtm}{document term matrix (the output from get_dtm function)}
}
\description{
This function takes the document term matrix (the output from get_dtm function) and builds a FIFA-themed word cloud of the top 100-most frequently used words
}
\examples{
buildMyFifaCloud(get_dtm(california_fifa, removeRare(scrubMe(california_fifa)))
}
