% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dtm.R
\name{get_dtm}
\alias{get_dtm}
\title{Create document term matrix}
\usage{
get_dtm(data, dict)
}
\arguments{
\item{data}{original dataframe of tweets}

\item{dict}{output from removeRare}
}
\value{
matrix
}
\description{
This function takes as inputs the original dataframe of tweets and the output from removeRare to create a document-term matrix
}
\examples{
get_dtm(california_fifa, removeRare(scrubMe(california_fifa)))
}
