% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyScore.R
\name{getMyScore}
\alias{getMyScore}
\title{Build sentiment scores from documents}
\usage{
getMyScore(words)
}
\arguments{
\item{words}{corpus of words to analyze}
}
\value{
numeric
}
\description{
This function calculates a sentiment score (number of positive word matches - number of negative word matches) from a collection of words. A positive sentiment score means more positive matches were detected than negative ones. A negative sentiment score means more negative matches were detected than positive ones. Special note: To get a sentiment score for each tweet (which could later be used to create a histogram of the frequency of sentiment scores), use sapply(a, b) where a is your output from scrubMe and b is the getMyScore function. To get a sentiment score for the whole collection of words/dictionary without paying attention to which tweet it corresponds to, just use getMyScore() where the input is the output from removeRare.
}
\examples{
getMyScore(removeRare(scrubMe(california_fifa)))
}
