% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWord.R
\name{plotWord}
\alias{plotWord}
\title{Map plot of all Tweets given keyword}
\usage{
plotWord(word)
}
\arguments{
\item{word}{string indicating keyword for plotting}
}
\description{
This function lets the user select a word and will then plot the tweets containing that word on a map of California with the size of each point depending on the number of retweets each tweet got. This function is dependent on the ggplot2 package.
}
\examples{
plotWord(word = "goal")
}
