% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWordSeries.R
\name{plotWordSeries}
\alias{plotWordSeries}
\title{Plot time series of word}
\usage{
plotWordSeries(word)
}
\arguments{
\item{word}{word to build time series}
}
\description{
A function that plots the time series of a particular word when it was tweeted in the California FIFA dataset. Input is the word you want to examine. Output is the time series by hour of the number of tweets containing that word.
}
\examples{
plotWordSeries("goal")
}
