% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRare.R
\name{removeRare}
\alias{removeRare}
\title{Remove rare words}
\usage{
removeRare(corpus)
}
\arguments{
\item{corpus}{Object of type “list” where each each row corresponds to a single tweet, and the words used in each tweet have been already been cleaned.}
}
\value{
vector
}
\description{
This function takes the output from the scrubMe function and gets rid of words that appeared less than 20 times. I.e., gets rid of the most rare words, which are kind of like outliers since they don’t represent how most people feel. They can also be words that are unintelligible.
}
\examples{
removeRare(scrubMe(california_fifa))
}
