% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempCheck.R
\name{tempCheck}
\alias{tempCheck}
\title{FIFA themed sentiment histogram}
\usage{
tempCheck(data)
}
\arguments{
\item{data}{raw dataframe collected using rtweet (column name of the text of your tweets must be called “full_text”)}
}
\description{
Quickly build a FIFA-themed histogram of sentiment scores without needing to do any prior pre-processing. Just input your raw dataframe collected using rtweet. Like above, the column name of the text of your tweets must be called “full_text”.
}
\examples{
tempCheck(california_fifa)
}
